# Have AttributesMap extend HashMap instead of contain LinkedHashMap