--- a/sdk/trace/src/main/java/io/opentelemetry/sdk/trace/AttributesMap.java
+++ b/sdk/trace/src/main/java/io/opentelemetry/sdk/trace/AttributesMap.java
@@ -9,7 +9,7 @@ import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.common.AttributesBuilder;
 import java.util.Collections;
-import java.util.LinkedHashMap;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.function.BiConsumer;
 
@@ -19,30 +19,24 @@ import java.util.function.BiConsumer;
  * <p>Note: this doesn't implement the Map interface, but behaves very similarly to one.
  */
 @SuppressWarnings({"rawtypes", "unchecked"})
-final class AttributesMap implements Attributes {
-  private final Map<AttributeKey<?>, Object> data;
+final class AttributesMap extends HashMap<AttributeKey<?>, Object> implements Attributes {
 
   private final long capacity;
   private int totalAddedValues = 0;
 
-  private AttributesMap(long capacity, Map<AttributeKey<?>, Object> data) {
-    this.capacity = capacity;
-    this.data = data;
-  }
-
   AttributesMap(long capacity) {
-    this(capacity, new LinkedHashMap<>());
+    this.capacity = capacity;
   }
 
-  public <T> void put(AttributeKey<T> key, T value) {
+  <T> void put(AttributeKey<T> key, T value) {
     if (key == null || key.getKey() == null || value == null) {
       return;
     }
     totalAddedValues++;
-    if (data.size() >= capacity && !data.containsKey(key)) {
+    if (size() >= capacity && !containsKey(key)) {
       return;
     }
-    data.put(key, value);
+    super.put(key, value);
   }
 
   int getTotalAddedValues() {
@@ -52,23 +46,13 @@ final class AttributesMap implements Attributes {
   @SuppressWarnings("unchecked")
   @Override
   public <T> T get(AttributeKey<T> key) {
-    return (T) data.get(key);
-  }
-
-  @Override
-  public int size() {
-    return data.size();
-  }
-
-  @Override
-  public boolean isEmpty() {
-    return data.isEmpty();
+    return (T) super.get(key);
   }
 
   @SuppressWarnings({"rawtypes", "unchecked"})
   @Override
-  public void forEach(BiConsumer<AttributeKey<?>, Object> consumer) {
-    for (Map.Entry<AttributeKey<?>, Object> entry : data.entrySet()) {
+  public void forEach(BiConsumer<? super AttributeKey<?>, ? super Object> consumer) {
+    for (Map.Entry<AttributeKey<?>, Object> entry : entrySet()) {
       AttributeKey key = entry.getKey();
       Object value = entry.getValue();
       consumer.accept(key, value);
@@ -77,7 +61,7 @@ final class AttributesMap implements Attributes {
 
   @Override
   public Map<AttributeKey<?>, Object> asMap() {
-    return Collections.unmodifiableMap(data);
+    return Collections.unmodifiableMap(this);
   }
 
   @Override
@@ -89,7 +73,7 @@ final class AttributesMap implements Attributes {
   public String toString() {
     return "AttributesMap{"
         + "data="
-        + data
+        + this
         + ", capacity="
         + capacity
         + ", totalAddedValues="
@@ -98,7 +82,6 @@ final class AttributesMap implements Attributes {
   }
 
   Attributes immutableCopy() {
-    Map<AttributeKey<?>, Object> dataCopy = new LinkedHashMap<>(data);
-    return new AttributesMap(capacity, Collections.unmodifiableMap(dataCopy));
+    return Attributes.builder().putAll(this).build();
   }
 }