--- a/sdk/trace/src/main/java/io/opentelemetry/sdk/trace/AttributesMap.java
+++ b/sdk/trace/src/main/java/io/opentelemetry/sdk/trace/AttributesMap.java
@@ -59,6 +59,11 @@ final class AttributesMap extends HashMap<AttributeKey<?>, Object> implements At
 
   @Override
   public Map<AttributeKey<?>, Object> asMap() {
+    // Because Attributes is marked Immutable, IDEs may recognize this as redundant usage. However,
+    // this class is private and is actually mutable, so we need to wrap with unmodifiableMap
+    // anyways. We implement the immutable Attributes for this class to support the
+    // Attributes.builder().putAll usage - it is tricky but an implementation detail of this private
+    // class.
     return Collections.unmodifiableMap(this);
   }