Contains #2905 

The ordering requirement of attributes was removed by @jkwatson so no need for `Linked`. Extension instead of composition (for this private class) saves us a few more bytes too, though we'd generally avoid that for public classes.

After
```
Benchmark                                                          Mode  Cnt     Score    Error   Units
FillSpanBenchmark.setFourAttributes                               thrpt   60  2106.729 ± 14.989  ops/ms
FillSpanBenchmark.setFourAttributes:·gc.alloc.rate                thrpt   60   791.987 ±  5.626  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.alloc.rate.norm           thrpt   60   592.000 ±  0.001    B/op
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Eden_Space       thrpt   60   793.685 ± 15.437  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Eden_Space.norm  thrpt   60   593.209 ± 10.138    B/op
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Old_Gen          thrpt   60     0.002 ±  0.001  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Old_Gen.norm     thrpt   60     0.002 ±  0.001    B/op
FillSpanBenchmark.setFourAttributes:·gc.count                     thrpt   60   472.000           counts
FillSpanBenchmark.setFourAttributes:·gc.time                      thrpt   60   180.000               ms
```

Before
```
Benchmark                                                          Mode  Cnt     Score    Error   Units
FillSpanBenchmark.setFourAttributes                               thrpt   60  1936.862 ± 18.334  ops/ms
FillSpanBenchmark.setFourAttributes:·gc.alloc.rate                thrpt   60   806.646 ±  7.654  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.alloc.rate.norm           thrpt   60   656.000 ±  0.001    B/op
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Eden_Space       thrpt   60   808.658 ± 15.555  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Eden_Space.norm  thrpt   60   657.776 ± 12.582    B/op
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Old_Gen          thrpt   60     0.002 ±  0.001  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Old_Gen.norm     thrpt   60     0.002 ±  0.001    B/op
FillSpanBenchmark.setFourAttributes:·gc.count                     thrpt   60   481.000           counts
FillSpanBenchmark.setFourAttributes:·gc.time                      thrpt   60   181.000               ms
```